fix: 바깥으로 나와있는 이미지파일과 관련 코드 수정
