+++ pubspec.yaml
@@ -60,15 +60,6 @@ flutter:
   # To add assets to your application, add an assets section, like this:
   assets:
     - assets/images/
-    - assets/logo.png
-    - assets/Vector.png
-    - assets/Vector2.png
-    - assets/pluse.png
-    - assets/p1.png
-    - assets/p2.png
-    - assets/p3.png
-    - assets/p4.png
-    - assets/p5.png
 
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/to/resolution-aware-images
